move minigames

+++ Gardener Witch Dungeon/Assets/Scripts/Enemy Skills/Basic Skeleton/BasicEnemyAttack.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class BasicEnemyAttack : EnemySkill
 {
-    [SerializeField] int damage;
+    [SerializeField] float damage;
     public override void Use()
     {
         EnemyManager.Instance.playerHealth.TakeDamage(damage);
